requirements/base.txt
# Code quality & testing
mypy
black
flake8
pytest
beautifulsoup4
lxml # Slightly faster html parser
django-stubs[compatible-mypy]

# Project
django
pytest-django
python-dotenv
markdown
pillow
types-Markdown
gunicorn